batch

+++ src/transformers/models/arcee/configuration_arcee.py
@@ -127,7 +127,7 @@ class ArceeConfig(PreTrainedConfig):
         hidden_act: str | None = "relu2",
         max_position_embeddings: int | None = 4096,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-5,
+        rms_norm_eps: float | None = 1e-5,
         use_cache: bool | None = True,
         pad_token_id: int | None = None,
         bos_token_id: int | None = 128000,
