+++ src/transformers/models/gemma/modular_gemma.py
@@ -145,7 +145,7 @@ class GemmaConfig(PreTrainedConfig):
         hidden_act: str | None = "gelu_pytorch_tanh",
         max_position_embeddings: int | None = 8192,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-6,
+        rms_norm_eps: float | None = 1e-6,
         use_cache: bool | None = True,
         pad_token_id: int | None = 0,
         eos_token_id: int | list[int] | None = 1,
