+++ src/transformers/models/minimax_m2/modular_minimax_m2.py
@@ -162,7 +162,7 @@ class MiniMaxM2Config(PreTrainedConfig):
         hidden_act: str | None = "silu",
         max_position_embeddings: int | None = 196608,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-06,
+        rms_norm_eps: float | None = 1e-06,
         use_cache: bool | None = True,
         pad_token_id: int | None = None,
         bos_token_id: int | None = 200034,
