+++ src/transformers/models/qwen2_5_omni/modular_qwen2_5_omni.py
@@ -357,7 +357,7 @@ class Qwen2_5OmniTextConfig(PreTrainedConfig):
         hidden_act: str | None = "silu",
         max_position_embeddings: int | None = 32768,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-6,
+        rms_norm_eps: float | None = 1e-6,
         use_cache: bool | None = True,
         rope_parameters: RopeParameters | dict[str, RopeParameters] | None = None,
         use_sliding_window: bool | None = False,
