+++ src/transformers/models/qwen2_5_vl/configuration_qwen2_5_vl.py
@@ -174,7 +174,7 @@ class Qwen2_5_VLTextConfig(PreTrainedConfig):
         hidden_act: str | None = "silu",
         max_position_embeddings: int | None = 32768,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-05,
+        rms_norm_eps: float | None = 1e-05,
         use_cache: bool | None = True,
         use_sliding_window: bool | None = False,
         sliding_window: int | None = 4096,
